irdeto: install cca lib and ca lib to casecure [1/1]
PD#SWPL-168226

Problem:
install cca lib and ca lib to casecure

Solution:
1, remove mode from irdeto-imw.bb
2, modify the aml hal source dir as aml_hal
3, When there is a CCA static lib in the SDK,
compile and install the CCA dynamic lib to casecure

Verify:
Yocto+BG201

+++ recipes-cas/irdeto/irdeto-sdk.bb
@@ -22,21 +22,45 @@ IRDETO_PATH:sc2 = "v3/dev/S905C2/signed"
 IRDETO_PATH:s4 = "s905c3"
 IRDETO_PATH:aq2432 = "s805c3"
 IRDETO_PATH:bf201 = "s805c1a"
+IRDETO_PATH:bg201 = "s805c1"
 
 REE_ONLY = "${@bb.utils.contains('DISTRO_FEATURES', 'irdeto-ree-only', 'true', 'false', d)}"
 
+EXTRA_OEMAKE = "STAGING_DIR=${STAGING_DIR_TARGET} TARGET_DIR=${D} \
+          IRDETO_PATH='${IRDETO_PATH}' \
+        "
+
+do_configure () {
+}
+
+do_compile () {
+    if [ -e ${S}/lib/irdeto/Makefile ] && [ -e ${S}/lib/irdeto/dvb/${IRDETO_PATH}/CloakedCAAgent.lib ]; then
+        cd ${S}/lib/irdeto/
+        oe_runmake -j1 ${EXTRA_OEMAKE} all
+    fi
+}
+
 do_install() {
-    install -d -m 0755 ${D}/usr/lib
     install -d -m 0755 ${D}/lib/optee_armtz
     install -d -m 0755 ${D}/usr/include
-    install -D -m 0644 ${S}/lib/ca/libirdetoca.so ${D}/usr/lib
+
     if [ "${REE_ONLY}" = "false" ]; then
         install -D -m 0644 ${S}/lib/ta/${IRDETO_PATH}/b64fd559-658d-48a4-bbc7-b95d8663f457.ta ${D}/lib/optee_armtz
         install -D -m 0644 ${S}/lib/ta/${IRDETO_PATH}/1d0f7170-3209-481c-b4ad-8fca95113b71.ta ${D}/lib/optee_armtz
     fi
+
+    if [ -e ${S}/lib/irdeto/out/ ]; then
+        install -d -m 0755 ${D}/casecure
+        cp ${S}/lib/irdeto/out/*.so ${D}/casecure
+        install -D -m 0644 ${S}/lib/ca/libirdetoca.so ${D}/casecure
+    else
+        install -d -m 0755 ${D}/usr/lib
+        install -D -m 0644 ${S}/lib/ca/libirdetoca.so ${D}/usr/lib
+    fi
     install -D -m 0644 ${S}/include/*.h ${D}/usr/include
 }
 
-FILES:${PN} = "${libdir}/* /lib/optee_armtz/*"
+SYSROOT_DIRS += "/casecure"
+FILES:${PN} = "${libdir}/* /lib/optee_armtz/* /casecure/* "
 FILES:${PN}-dev = "${includedir}/* "
 INSANE_SKIP:${PN} = "dev-so ldflags dev-elf"
